requirements.txt
django==1.8.3
django-cachalot==2.4.0
werkzeug==2.3.7 
idna==2.10 
requests==2.24.0 
scikit-learn==0.20.1 
huggingface-llm-test==0.1.0
transformers==4.57.1 
setuptools==65.5.0 
flask==3.1.2 